Add meta file to check if downloaded

+++ xinference/core/gradio.py
@@ -266,12 +266,17 @@ class GradioApp:
             progress=gr.Progress(),
         ):
             model_family = MODEL_TO_FAMILIES[_model_name]
-            cache_path = model_family.generate_cache_path(
+            cache_path, meta_path = model_family.generate_cache_path(
                 int(_model_size_in_billions), _quantization
             )
-            if not os.path.exists(cache_path):
+            if not (os.path.exists(cache_path) and os.path.exists(meta_path)):
+                if os.path.exists(cache_path):
+                    os.remove(cache_path)
                 url = model_family.url_generator(
                     int(_model_size_in_billions), _quantization
+                )
+                full_name = (
+                    f"{str(model_family)}-{_model_size_in_billions}b-{_quantization}"
                 )
                 try:
                     urllib.request.urlretrieve(
@@ -282,6 +287,9 @@ class GradioApp:
                             desc=self._locale("Downloading"),
                         ),
                     )
+                    # write a meta file to record if download finished
+                    with open(meta_path, "w") as f:
+                        f.write(full_name)
                 except:
                     if os.path.exists(cache_path):
                         os.remove(cache_path)
